filepath_is_ancestor(): fix handling of cases where wd2 != wd1
If in a rename() syscall the new path name is relative, and starts
from a (direct or indirect) child of the old path,
filepath_is_ancestor() fails to detect the ancestor relationship,
because the tuple corresponding to the old path is never
encountered when traversing the new path; as a result, if a rename
tries to to make a directory a subdirectory of itself, the kernel
crashes in an infinite recursion loop.
This commit fixes the above issue by traversing the new path from
the innermost child to the root tuple, and checking at every step
whether there is a match with the old path.
A test case that would have exposed the above issue is being added
to the "rename" runtime tests.