Implement protection against arbitrary code execution
This commit implements an "exec protection" feature that, if
enabled, prevents the kernel from executing any code outside the
main program executable and other "trusted" files explicitly marked
as executable. This feature is enabled by adding an
"exec_protection" symbol to the root tuple in the manifest file;
then, any files containing executable code that need to be executed
by the application (usually, shared libraries) need to marked in
the manifest file with an "exec" symbol. It is not necessary to
add the "exec" symbol to the main program, since this is always
considered as a "trusted" executable file.
Example of a manifest file with exec protection enabled and with
one shared library:
(
  children:(
    my_app:(contents:(host:bin/my_app))
    lib64:(children:(ld-linux-x86-64.so.2:(contents:(host:/lib64/ld-linux-x86-64.so.2))))
    lib:(children:(x86_64-linux-gnu:(children:(libc.so.6:(exec:t contents:(host:/lib/x86_64-linux-gnu/libc.so.6))))))
  )
    program:/my_app
    environment:(USER:bobby PWD:/)
    exec_protection:t
)
(Note the exec symbol in the libc.so.6 shared library, without
which the application would fail to start, with an error such as
"error while loading shared libraries: libc.so.6: failed to map
segment from shared object".)